+++ kernel/ideals.cc
@@ -768,7 +768,7 @@ ideal idLiftStd (ideal  h1, matrix* ma, tHomog hi, ideal * syz)
   BITSET save2;
   SI_SAVE_OPT2(save2);
 
-  k=si_max((long)1,id_RankFreeModule(h1,currRing));
+  k=si_max(1L,id_RankFreeModule(h1,currRing));
 
   if ((k==1) && (!lift3)) si_opt_2 |=Sy_bit(V_IDLIFT);
